Refactor Duplicator
