+++ data/loop.gfa
@@ -1,0 +1,7 @@
+H	VN:Z:1.0
+S	read1	GATTACA
+S	read2	ACAGATTACA
+S	read3	ACAGAGG
+L	read1	+	read2	+	3M
+L	read2	+	read3	+	3M
+L	read2	+	read2	+	3M
